[python] Prevent illegal svg id characters
If a user input some special chars (like 'A&B') in a party name, the
python used to create a <g id='A&B'> element, which is illegal. It may
now create several <g> with the same id, but this issue way already
present if the user used 2 parties with the same name.