newarch: revamp
Calculations are remade almost from scratch, the totals calculations are remade too, the dense_rows option is reimplemented...
This is what went into the Revamp branch, basically.
This will all be made obsolete by the new parliamentarch python module, but it induces some changes to the generated diagrams - which I find to be better but are yet material changes.